requirements.txt
Django==3.0.8
django-crispy-forms==1.9.2
Pillow==7.2.0
django-filter
django-bootstrap-form